refactor and add testing
