+++ cycletls/go.mod
@@ -4,10 +4,11 @@ go 1.14
 
 require (
 	github.com/Danny-Dasilva/fhttp v0.0.0-20211010093114-56fde831fe2f
+	github.com/andybalholm/brotli v1.0.3
 	github.com/gorilla/websocket v1.4.2
 	gitlab.com/yawning/utls.git v0.0.12-1
 	golang.org/x/crypto v0.0.0-20210921155107-089bfa567519 // indirect
-	golang.org/x/net v0.0.0-20211005215030-d2e5035098b3
+	golang.org/x/net v0.0.0-20211008194852-3b03d305991f
 	golang.org/x/sys v0.0.0-20211004093028-2c5d950f24ef // indirect
 	golang.org/x/text v0.3.7 // indirect
 )
