+++ cycletls/roundtripper.go
@@ -2,13 +2,11 @@ package cycletls
 
 import (
 	"context"
-	"crypto/sha256"
 	"crypto/tls"
 	"errors"
 	"fmt"
 	"net"
 
-	"strconv"
 	"strings"
 	"sync"
 
@@ -194,157 +192,4 @@ func newRoundTripper(browser browser, dialer ...proxy.ContextDialer) http.RoundT
 		cachedTransports:  make(map[string]http.RoundTripper),
 		cachedConnections: make(map[string]net.Conn),
 	}
-}
-
-///////////////////////// test code
-
-// StringToSpec creates a ClientHelloSpec based on a JA3 string
-func StringToSpec(ja3 string) (*utls.ClientHelloSpec, error) {
-	extMap := genMap()
-	tokens := strings.Split(ja3, ",")
-
-	version := tokens[0]
-	ciphers := strings.Split(tokens[1], "-")
-	extensions := strings.Split(tokens[2], "-")
-	curves := strings.Split(tokens[3], "-")
-	if len(curves) == 1 && curves[0] == "" {
-		curves = []string{}
-	}
-	pointFormats := strings.Split(tokens[4], "-")
-	if len(pointFormats) == 1 && pointFormats[0] == "" {
-		pointFormats = []string{}
-	}
-
-	// parse curves
-	var targetCurves []utls.CurveID
-	targetCurves = append(targetCurves, utls.CurveID(utls.CurveID(utls.GREASE_PLACEHOLDER))) //append grease for Chrome browsers
-	for _, c := range curves {
-		cid, err := strconv.ParseUint(c, 10, 16)
-		if err != nil {
-			return nil, err
-		}
-		targetCurves = append(targetCurves, utls.CurveID(cid))
-	}
-	extMap["10"] = &utls.SupportedCurvesExtension{Curves: targetCurves}
-
-	// parse point formats
-	var targetPointFormats []byte
-	for _, p := range pointFormats {
-		pid, err := strconv.ParseUint(p, 10, 8)
-		if err != nil {
-			return nil, err
-		}
-		targetPointFormats = append(targetPointFormats, byte(pid))
-	}
-	extMap["11"] = &utls.SupportedPointsExtension{SupportedPoints: targetPointFormats}
-
-	// set extension 43
-	vid64, err := strconv.ParseUint(version, 10, 16)
-	if err != nil {
-		return nil, err
-	}
-	vid := uint16(vid64)
-	// extMap["43"] = &utls.SupportedVersionsExtension{
-	// 	Versions: []uint16{
-	// 		utls.VersionTLS12,
-	// 	},
-	// }
-
-	// build extenions list
-	var exts []utls.TLSExtension
-	for _, e := range extensions {
-		te, ok := extMap[e]
-		if !ok {
-			return nil, raiseExtensionError(e)
-		}
-		exts = append(exts, te)
-	}
-	// build SSLVersion
-	// vid64, err := strconv.ParseUint(version, 10, 16)
-	// if err != nil {
-	// 	return nil, err
-	// }
-
-	// build CipherSuites
-	var suites []uint16
-	for _, c := range ciphers {
-		cid, err := strconv.ParseUint(c, 10, 16)
-		if err != nil {
-			return nil, err
-		}
-		suites = append(suites, uint16(cid))
-	}
-	_ = vid
-	return &utls.ClientHelloSpec{
-		// TLSVersMin:         vid,
-		// TLSVersMax:         vid,
-		CipherSuites:       suites,
-		CompressionMethods: []byte{0},
-		Extensions:         exts,
-		GetSessionID:       sha256.Sum256,
-	}, nil
-}
-
-func genMap() (extMap map[string]utls.TLSExtension) {
-	extMap = map[string]utls.TLSExtension{
-		"0": &utls.SNIExtension{},
-		"5": &utls.StatusRequestExtension{},
-		// These are applied later
-		// "10": &tls.SupportedCurvesExtension{...}
-		// "11": &tls.SupportedPointsExtension{...}
-		"13": &utls.SignatureAlgorithmsExtension{
-			SupportedSignatureAlgorithms: []utls.SignatureScheme{
-				utls.ECDSAWithP256AndSHA256,
-				utls.ECDSAWithP384AndSHA384,
-				utls.ECDSAWithP521AndSHA512,
-				utls.PSSWithSHA256,
-				utls.PSSWithSHA384,
-				utls.PSSWithSHA512,
-				utls.PKCS1WithSHA256,
-				utls.PKCS1WithSHA384,
-				utls.PKCS1WithSHA512,
-				utls.ECDSAWithSHA1,
-				utls.PKCS1WithSHA1,
-			},
-		},
-		"16": &utls.ALPNExtension{
-			AlpnProtocols: []string{"h2", "http/1.1"},
-		},
-		"18": &utls.SCTExtension{},
-		"21": &utls.UtlsPaddingExtension{GetPaddingLen: utls.BoringPaddingStyle},
-		"22": &utls.GenericExtension{Id: 22}, // encrypt_then_mac
-		"23": &utls.UtlsExtendedMasterSecretExtension{},
-		"27": &utls.CompressCertificateExtension{
-			Algorithms: []utls.CertCompressionAlgo{utls.CertCompressionBrotli},
-		},
-		"28": &utls.FakeRecordSizeLimitExtension{}, //Limit: 0x4001
-		"35": &utls.SessionTicketExtension{},
-		"34": &utls.GenericExtension{Id: 34},
-		"41": &utls.GenericExtension{Id: 41}, //FIXME pre_shared_key
-		"43": &utls.SupportedVersionsExtension{Versions: []uint16{
-			utls.GREASE_PLACEHOLDER,
-			utls.VersionTLS13,
-			utls.VersionTLS12,
-			utls.VersionTLS11,
-			utls.VersionTLS10}},
-		"44": &utls.CookieExtension{},
-		"45": &utls.PSKKeyExchangeModesExtension{Modes: []uint8{
-			utls.PskModeDHE,
-		}},
-		"49": &utls.GenericExtension{Id: 49}, // post_handshake_auth
-		"50": &utls.GenericExtension{Id: 50}, // signature_algorithms_cert
-		"51": &utls.KeyShareExtension{KeyShares: []utls.KeyShare{
-			{Group: utls.CurveID(utls.GREASE_PLACEHOLDER), Data: []byte{0}},
-			{Group: utls.X25519},
-
-			// {Group: utls.CurveP384}, known bug missing correct extensions for handshake
-		}},
-		"30032": &utls.GenericExtension{Id: 0x7550, Data: []byte{0}}, //FIXME
-		"13172": &utls.NPNExtension{},
-		"65281": &utls.RenegotiationInfoExtension{
-			Renegotiation: utls.RenegotiateOnceAsClient,
-		},
-	}
-	return
-
 }
